train-progress/my_website/live2d/requirements.txt
fastapi>=0.100.0
pydantic>=2.0.0
openai>=1.0.0
python-dotenv>=1.0.0
uvicorn>=0.23.0
requests>=2.31.0
httpx>=0.24.0
